requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
python-dotenv==1.0.0
Pillow==10.1.0
pytesseract==0.3.13
opencv-python-headless==4.7.0.72
easyocr==1.7.1
google-cloud-vision==3.7.2
numpy==1.26.4



